+++ sdk/python/pulumi_fastly/__init__.py
@@ -4,11 +4,13 @@
 
 # Export this package's modules as members:
 from .get_fastly_ip_ranges import *
+from .get_waf_rules import *
 from .provider import *
 from .service_acl_entriesv1 import *
 from .service_compute import *
 from .service_dictionary_itemsv1 import *
 from .service_dynamic_snippet_contentv1 import *
+from .service_waf_configuration import *
 from .servicev1 import *
 from .userv1 import *
 from ._inputs import *